Momentum + control changes (EXPERIMENTAL)
NOTE: I was unable to figure out how to run the game -- the play button
in Unity wasn't doing anything, for whatever reason. DO NOT ASSUME THIS
CODE WILL WORK BEFORE TESTING IT. In particular, Q and E may be
reversed.

Q and E should rotate the view horizontally, R rises, F descends.

No vertical view rotation is implemented, so we may need to figure out
what keys we want to use for that, if any? (Since rising/descending
isn't based on which way you're facing, though, we should be able to
just put everything "within view" and not worry about it.)