Title Screen project + graphics
Currently the button "sticks" in highlighted mode when pressed. This is
a known issue with Unity, and since it's 1) the only button we have and
2) the button to start the game, it's probably not a huge issue.

Not sure how scene transitions in Unity work; if I have time, I'll look
this over after I take care of composing + other stuff. If not, we'll
get the button properly functional in class or over the weekend.